backend/library/requirements.txt
Django==3.2.5
djangorestframework
markdown
django-filter